+++ NEWS.md
@@ -6,7 +6,12 @@
 * Release date: 2025-04-22 (April 22, 2025)
 * Public release date: 2025-05-02 (May 2, 2025)
 
+## Version 0.0.2
 
+* Added parametric modeling with parfitml and changed the name of the
+  estimSI routine to nonparfit.
+* Version name: "Parametric wave 1".
+* Release date: 2025-12-14 (December 14, 2025)